requirements.txt
numpy
SimpleITK
pydicom
pydicom-seg
nibabel
dcmstack
pyradiomics
tcia_utils
inquirer
tqdm
python-dotenv